Convert rows and cols variables to size_t
There are a number of places where rows and cols being "large"
aren't supported, but at least make the storage for them be
size_t so that serialization is future-proof.

+++ GM/DlgBrdsz.h
@@ -41,8 +41,8 @@ public:
     CEdit   m_editCellHt;
     UINT    m_nCellHt;
     UINT    m_nCellWd;
-    UINT    m_nCols;
-    UINT    m_nRows;
+    size_t  m_nCols;
+    size_t  m_nRows;
     BOOL    m_bStaggerIn;
     //}}AFX_DATA
